Seperate function uploadPhotos from my-map page
